displaye joke

+++ app/admin/component/edit-form.tsx
@@ -1,5 +1,6 @@
 import { JokeContext } from "@/context/joke-context";
 import React, { useContext } from "react";
+import { ClipLoader } from "react-spinners";
 
 export const jokeTypes = [
   "Any",
@@ -22,12 +23,36 @@ export default function UpdateJoke() {
     setSelectedJokeType,
     selectedJokeId,
     deleteJoke,
+    selectedJokeData,
+    submitJoke,
+    isDeletingLoading,
+    isSubmittingLoading,
   } = useContext(JokeContext);
 
   const handleDelete = (e: any) => {
-    e.preventDefault(); // Prevent form submission
-    if (selectedJokeId) {
-      deleteJoke(selectedJokeId);
+    e.preventDefault();
+
+    // Ask for confirmation before deleting
+    if (window.confirm("Are you sure you want to delete this joke?")) {
+      if (selectedJokeId) {
+        deleteJoke(selectedJokeId);
+      }
+    } else {
+      return;
+    }
+  };
+
+  const handleSubmit = (e: any) => {
+    e.preventDefault();
+
+    // Ask for confirmation
+    if (window.confirm("Are you sure you want to submit?")) {
+      if (selectedJokeData) {
+        selectedJokeData.type = selectedJokeType || "";
+        submitJoke(selectedJokeData);
+      }
+    } else {
+      return;
     }
   };
 
@@ -48,7 +73,7 @@ export default function UpdateJoke() {
         <select
           name="type"
           id="type"
-          value={selectedJokeType || ""} // Ensure a default empty string
+          value={selectedJokeType || ""}
           onChange={(e) => setSelectedJokeType(e.target.value)}
           className="border border-black rounded-md p-2"
         >
@@ -60,14 +85,21 @@ export default function UpdateJoke() {
         </select>
       </div>
       <div className="w-full space-x-1 flex">
-        <button className="bg-purple-600 text-white rounded-md p-2 my-5 w-1/2 justify-center">
-          Update
+        <button
+          className="bg-red-600 text-white rounded-md p-2 my-5 w-1/3 justify-center"
+          onClick={handleDelete}
+          disabled={isDeletingLoading || selectedJokeId === null}
+        >
+          <ClipLoader color="white" loading={isDeletingLoading} size={20} />
+          Remove
         </button>
         <button
-          className="bg-red-600 text-white rounded-md p-2 my-5 w-1/2 justify-center"
-          onClick={handleDelete}
+          className="bg-green-600 text-white rounded-md p-2 my-5 w-2/3 justify-center"
+          onClick={handleSubmit}
+          disabled={isSubmittingLoading || selectedJokeId === null}
         >
-          Remove
+          <ClipLoader color="white" loading={isSubmittingLoading} size={20} />
+          Submit
         </button>
       </div>
     </form>
